requirements.txt
# Core dependencies
gigachat==0.1.38
pydantic==2.10.6
python-dotenv==1.0.1
loguru==0.7.3

# Database dependencies
psycopg2-binary==2.9.9

# Kafka dependencies
aiokafka==0.8.1
kafka-python==2.0.2
six==1.16.0

# Testing dependencies
pytest==8.3.5
httpx==0.27.2
allure-pytest==2.13.2

# Required by dependencies (do not remove)
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
h11==0.14.0
httpcore==1.0.7
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic-core==2.27.2
sniffio==1.3.1
starlette==0.46.0
typing-extensions==4.12.2
